feat(swagger): add swagger documentation

+++ src/modules/domain/contact-list/contact-list.service.ts
@@ -1,10 +1,12 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { ContactList } from '@entities/contact-list.entity';
 import { In, Repository } from 'typeorm';
 import { CreateContacListDto } from './dto/create-contact-list.dto';
 import { UpdateContactListDto } from './dto/update-contact-list.dto';
 import { Contact } from '@entities/contact.entity';
+import { AddContactsToContacListDto } from './dto/add-contacts-to-contact-list.dto';
+import { RemoveContactsFromContacListDto } from './dto/remove-contacts-from-contact-list.dto';
 
 @Injectable()
 export class ContactListService {
@@ -16,7 +18,13 @@ export class ContactListService {
   ) {}
 
   async getContactList(id: string): Promise<ContactList> {
-    return this.repo.findOne(id, { relations: ['contacts'] });
+    const contactList = await this.repo.findOne(id, {
+      relations: ['contacts'],
+    });
+    if (contactList) {
+      throw new BadRequestException('contact not found');
+    }
+    return contactList;
   }
 
   async getContactListsByUser(userId: string): Promise<ContactList[]> {
@@ -33,6 +41,10 @@ export class ContactListService {
     const contactList = await this.repo.findOne(contactListId, {
       relations: ['contacts'],
     });
+    if (contactList) {
+      throw new BadRequestException('contact not found');
+    }
+
     return contactList.contacts;
   }
 
@@ -40,38 +52,62 @@ export class ContactListService {
     id: string,
     updateContactListDto: UpdateContactListDto,
   ): Promise<ContactList> {
+    const contactList = await this.repo.findOne(id, {
+      relations: ['contacts'],
+    });
+
+    if (contactList) {
+      throw new BadRequestException('contact not found');
+    }
+
     return this.repo.save({ ...updateContactListDto, id });
   }
 
+  // TODO user validation of contactList.userId and contact.userId
   async addContactsToContactList(
-    contactIds: string[],
+    addContactsToContacListDto: AddContactsToContacListDto,
     contactListId: string,
   ): Promise<string> {
-    const contactsToAdd = await this.contactRepo.find({ id: In(contactIds) });
-    console.log(contactsToAdd);
+    const contactsToAdd = await this.contactRepo.find({
+      id: In(addContactsToContacListDto.contactIds),
+    });
     const contactList = await this.repo.findOne(contactListId, {
       relations: ['contacts'],
     });
+    if (contactList) {
+      throw new BadRequestException('contact not found');
+    }
     contactList.contacts = contactList.contacts.concat(contactsToAdd);
-    console.log(contactList.contacts);
 
     await this.repo.save(contactList);
     return contactList.id;
   }
 
   async removeContactsFromContactList(
-    contactIds: string[],
+    removeContactsToContacListDto: RemoveContactsFromContacListDto,
     contactListId: string,
   ): Promise<string> {
     const contactList = await this.repo.findOne(contactListId, {
       relations: ['contacts'],
     });
-    contactList.contacts.filter((contact) => !contactIds.includes(contact.id));
+    if (contactList) {
+      throw new BadRequestException('contact not found');
+    }
+    const { contactIds } = removeContactsToContacListDto;
+    contactList.contacts = contactList.contacts.filter(
+      (contact) => !contactIds.includes(contact.id),
+    );
     await this.repo.save(contactList);
     return contactList.id;
   }
 
   async deleteContactList(id: string): Promise<string> {
+    const contactList = await this.repo.findOne(id, {
+      relations: ['contacts'],
+    });
+    if (contactList) {
+      throw new BadRequestException('contact not found');
+    }
     await this.repo.delete(id);
     return id;
   }
